cleanup, add correct challenge id
delete unused

cleanup

cleanup
