+++ packages/dapp/src/components/listinglist/ListingListItem.tsx
@@ -9,12 +9,10 @@ import { getFormattedTokenBalance } from "@joincivil/utils";
 import {
   ListingContainerProps,
   connectLatestChallengeSucceededResults,
-  // connectGraphQLLatestChallengeSucceededResults,
   getChallengeResultsProps,
 } from "../utility/HigherOrderComponents";
 import WhitelistedListingItem from "./WhitelistedListingItem";
 import { getContent, getBareContent } from "../../redux/actionCreators/newsrooms";
-import BigNumber from "bignumber.js";
 
 export interface ListingListItemOwnProps {
   listingAddress?: string;
@@ -65,7 +63,6 @@ class ListingListItem extends React.Component<ListingListItemOwnProps & ListingL
   }
 
   public render(): JSX.Element {
-    console.log("renderListing. props: ", this.props);
     const { listing, newsroom, listingPhaseState } = this.props;
     const listingExists = listing && listing.data && newsroom && listingPhaseState;
     const isWhitelisted = listingExists && listingPhaseState.isWhitelisted;
@@ -152,7 +149,6 @@ class ListingListItem extends React.Component<ListingListItemOwnProps & ListingL
 
 const RejectedListing: React.StatelessComponent<ListingListItemOwnProps & ListingListItemReduxProps> = props => {
   const { listingAddress, newsroom, listingPhaseState, charter, listing } = props;
-  console.log("rejected listing props: ", props);
   const newsroomData = newsroom!.data;
   const listingDetailURL = `/listing/${listingAddress}`;
   let description = "";
@@ -168,25 +164,26 @@ const RejectedListing: React.StatelessComponent<ListingListItemOwnProps & Listin
     listingDetailURL,
     ...listingPhaseState,
   };
-  let ListingSummaryRejected;
-  if (!listing!.data.prevChallenge) {
-    ListingSummaryRejected = compose<React.ComponentClass<ListingContainerProps & {}>>(
+  const data = listing!.data!;
+  if (!data.prevChallenge) {
+    const ListingSummaryRejected = compose<React.ComponentClass<ListingContainerProps & {}>>(
       connectLatestChallengeSucceededResults,
     )(ListingSummaryRejectedComponent);
     return <ListingSummaryRejected {...listingViewProps} />;
   } else {
-    // ListingSummaryRejected = compose<React.ComponentClass<ListingContainerProps & {}>>(
-    //   connectGraphQLLatestChallengeSucceededResults
-    // )(ListingSummaryRejectedComponent);
     const challengeResultsProps = getChallengeResultsProps({
-      listingAddress: "0",
-      challengeID: new BigNumber("1"),
-      challenge: listing!.data.prevChallenge!,
+      listingAddress: listingAddress!,
+      challengeID: data.prevChallengeID!,
+      challenge: data.prevChallenge!,
     });
-    return <ListingSummaryRejectedComponent challengeID="1" {...challengeResultsProps} {...listingViewProps} />;
+    return (
+      <ListingSummaryRejectedComponent
+        challengeID={data.prevChallengeID!.toString()}
+        {...challengeResultsProps}
+        {...listingViewProps}
+      />
+    );
   }
-  // console.log("ListingSummaryRejected: ", ListingSummaryRejected);
-  // return <ListingSummaryRejected {...listingViewProps} />;
 };
 
 const mapStateToProps = (
